Updated formatting and fixed include statements in both kernels. Moved eventconfiguration struct to seperate file.
